fix values type

+++ main.py
@@ -1,6 +1,8 @@
 from fastapi import FastAPI, HTTPException, Depends
 from sqlalchemy.orm import Session
 import uuid
+from decimal import Decimal
+
 from models.models import OperationType, Wallet, Operation
 from database import get_db
 
@@ -14,11 +16,11 @@ def modify_balance(wallet_id: uuid.UUID, operation: Operation, db: Session = Dep
             raise HTTPException(status_code=404, detail="Wallet not found")
 
         if operation.operation_type == OperationType.DEPOSIT:
-            wallet.balance = float(wallet.balance) + operation.amount
+            wallet.balance = Decimal(wallet.balance) + operation.amount
         elif operation.operation_type == OperationType.WITHDRAW:
             if wallet.balance < operation.amount:
                 raise HTTPException(status_code=400, detail="Insufficient funds")
-            wallet.balance = float(wallet.balance) - operation.amount
+            wallet.balance = Decimal(wallet.balance) - operation.amount
         db.commit()
     return {"wallet_id": str(wallet_id), "new_balance": wallet.balance}
